Solving print, changing double to int and adding Johnson's Algorithm

+++ src/main/java/com/biedermann/pathfindingalgorithms/Services/MatrixReader.java
@@ -10,17 +10,17 @@ import java.util.List;
 
 @Service
 public class MatrixReader {
-    public double[][] readAdjacencyMatrix(String filePath) throws IOException {
+    public int[][] readAdjacencyMatrix(String filePath) throws IOException {
         Path path = Paths.get(filePath);
         List<String> lines = Files.readAllLines(path);
 
         int size = Integer.parseInt(lines.get(0));
-        double[][] adjacencyMatrix = new double[size][size];
+        int[][] adjacencyMatrix = new int[size][size];
 
         for (int i = 0; i < size; i++) {
             String [] numbers = lines.get(i + 1).split(" ");
             for (int j = 0; j < size; j++) {
-                adjacencyMatrix[i][j] = numbers[j].equalsIgnoreCase("INF") ? Double.MAX_VALUE : Double.parseDouble(numbers[j]);
+                adjacencyMatrix[i][j] = numbers[j].equalsIgnoreCase("INF") ? Integer.MAX_VALUE : Integer.parseInt(numbers[j]);
             }
         }
         return adjacencyMatrix;
